+++ src/components/formInjector/formInjector.component.ts
@@ -3,31 +3,18 @@ import {CheckboxFieldDataInterface} from '../../inputs/checkbox/checkbox.interfa
 import {Component, Input} from '@angular/core'
 import {DatepickerFieldDataInterface} from '../../inputs/datepicker/datepicker.interfaces'
 import {FileInputFieldDataInterface} from '../../inputs/file/file.interfaces'
+import {FormFieldsInterface, FormLayoutColumnDataInterface} from '../../services/formBuilder'
 import {InputFieldDataInterface} from '../../inputs/input/input.interfaces'
 import {SelectFieldDataInterface} from '../../inputs/select/select.interfaces'
 import {SlideToggleFieldDataInterface} from '../../inputs/slideToggle/slideToggle.interfaces'
 import {TextareaFieldDataInterface} from '../../inputs/textarea/textarea.interfaces'
 
 @Component({
-	selector: 'rui-input-injector',
-	templateUrl: './inputInjector.template.html'
+	selector: 'rui-form-injector',
+	templateUrl: './formInjector.template.html'
 })
-export class InputInjector {
-	regularInputTypes = [
-		'color',
-		'email',
-		'month',
-		'number',
-		'password',
-		'search',
-		'tel',
-		'text',
-		'time',
-		'url',
-		'week'
-	]
-
-	@Input() type: string
+export class FormInjector {
+	@Input() fieldConfig: FormFieldsInterface[]
 	@Input() fieldData:
 		AutocompleteFieldDataInterface |
 		CheckboxFieldDataInterface |
@@ -37,6 +24,8 @@ export class InputInjector {
 		SelectFieldDataInterface |
 		SlideToggleFieldDataInterface |
 		TextareaFieldDataInterface
+	@Input() layout: FormLayoutColumnDataInterface[][]
+	@Input() rowOffsets: number[] = []
 
 	constructor() {
 	}