+++ src/services/formBuilder/formBuilder.interfaces.ts
@@ -2,10 +2,27 @@ import {AutocompleteFieldDataInterface} from '../../inputs/autocomplete/autocomp
 import {CheckboxFieldDataInterface} from '../../inputs/checkbox/checkbox.interfaces'
 import {DatepickerFieldDataInterface} from '../../inputs/datepicker/datepicker.interfaces'
 import {FileInputFieldDataInterface} from '../../inputs/file/file.interfaces'
+import {FormGroup} from '@angular/forms'
 import {InputFieldDataInterface} from '../../inputs/input/input.interfaces'
 import {SelectFieldDataInterface} from '../../inputs/select/select.interfaces'
 import {SlideToggleFieldDataInterface} from '../../inputs/slideToggle/slideToggle.interfaces'
 import {TextareaFieldDataInterface} from '../../inputs/textarea/textarea.interfaces'
+
+export interface BuildFormReturnDataInterface {
+	form: FormGroup
+	fieldData: {
+		[fieldName: string]:
+			AutocompleteFieldDataInterface |
+			CheckboxFieldDataInterface |
+			DatepickerFieldDataInterface |
+			FileInputFieldDataInterface |
+			InputFieldDataInterface |
+			SelectFieldDataInterface |
+			SlideToggleFieldDataInterface |
+			TextareaFieldDataInterface
+	}
+	layout?: FormLayoutColumnDataInterface[][]
+}
 
 export interface FormFieldsInterface {
 	autocompleteConfig?: AutocompleteFieldDataInterface
@@ -18,6 +35,12 @@ export interface FormFieldsInterface {
 	label: string
 	masterFieldName?: string
 	name: string
+	positioning?: {
+		colOffset?: string
+		colSize: string
+		rowIndex: number
+		rowSpan?: number
+	}
 	selectConfig?: SelectFieldDataInterface
 	slideToggleConfig?: SlideToggleFieldDataInterface
 	textareaConfig?: TextareaFieldDataInterface
@@ -35,7 +58,6 @@ export interface FormFieldsInterface {
 		'search' |
 		'tel' |
 		'text' |
-		'textarea' |
 		'time' |
 		'url' |
 		'week' |
@@ -57,4 +79,12 @@ export interface FormFieldValidationsInterface {
 		'required' |
 		'validateEmail'
 	value?: any
+}
+
+export interface FormLayoutColumnDataInterface {
+	colOffset?: string
+	colSize?: string
+	fieldName?: string
+	innerRows?: FormLayoutColumnDataInterface[][]
+	itemIndex?: number
 }